lt - move tag parser into services directory
