+++ app/helpers/images_helper.rb
@@ -1,12 +1,2 @@
 module ImagesHelper
-  class ImageTagParser < ActsAsTaggableOn::GenericParser
-    def parse
-      ActsAsTaggableOn::TagList.new.tap do |tag_list|
-        tag_list.add @tag_list.split(' ')
-        tag_list.each do |tag|
-          tag.prepend('#') unless tag.starts_with? '#' # make it a hash tag if it isn't already
-        end
-      end
-    end
-  end
 end
